requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
flask>=2.3.0
flask-cors>=4.0.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
openai>=1.0.0
pillow>=10.0.0
pytesseract>=0.3.10
